WIP: 1. implement parse_cad_sch() in start_bag.il; 2. move geenric methods from oa.py and skill.py into database.py; 3. implement _import_design() in skill.py
